+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Status: Available for use
 ## [Unreleased]
 ### Changed
 - Cleanup of resolution of working directory to use proper path types - PATCH
+- Cleanup and improve handling of user uid and gid - PATCH
 
 ### Added